+++ GraduationTracker/GraduationTracker/GraduationTracker.csproj
@@ -45,7 +45,7 @@
     <Compile Include="GraduationTracker.cs" />
     <Compile Include="STANDING.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Repository.cs" />
+    <Compile Include="Repositories\Repository.cs" />
     <Compile Include="Requirement.cs" />
     <Compile Include="Student.cs" />
   </ItemGroup>
